Cleanup and Major Functionality changes.
Changed a QLabel from, "label" to "IWAD_label" so its more descriptive. Cleaned up a lot of my code and comments, redid how some functionality works.

When the Turok engine is chosen, non-applicable options are disabled, and on engine launch, are never used. As of right now, this is fully Turok compatible, but quality of life tweaks will be added.