UI Tweaks + Launch without Files/Archives
I've cleaned up the UI and set it back to a grid. The engine tab has been converted to a horizontal layout that in my opinion looks nicer.

I've added the ability to launch games wihout any files/archives or PWADs. Useful for quick vanilla play or development of mods.